+++ lib/dealog_backoffice/router.ex
@@ -12,6 +12,7 @@ defmodule DealogBackoffice.Router do
     CreateMessage,
     ChangeMessage,
     SendMessageForApproval,
+    ApproveMessage,
     RejectMessage
   }
 
@@ -21,5 +22,5 @@ defmodule DealogBackoffice.Router do
 
   identify(Message, by: :message_id, prefix: "message-")
 
-  dispatch([CreateMessage, ChangeMessage, SendMessageForApproval, RejectMessage], to: Message)
+  dispatch([CreateMessage, ChangeMessage, SendMessageForApproval, ApproveMessage, RejectMessage], to: Message)
 end