+++ test/dealog_backoffice_web/live/organization_messages/edit_test.exs
@@ -1,4 +1,4 @@
-defmodule DealogBackofficeWeb.OrganizationMessagesLive.MessageLiveTest do
+defmodule DealogBackofficeWeb.OrganizationMessagesLive.EditTest do
   use DealogBackofficeWeb.ConnCase
 
   import Phoenix.LiveViewTest